backbone/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
django-cors-headers==3.1.0
django-filter==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.10.2
djangorestframework-jsonapi==2.8.0
djangorestframework-simplejwt==4.3.0
idna==2.8
inflection==0.3.1
mysqlclient==1.4.4
PyJWT==1.7.1
PyPDF2==1.26.0
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
